+++ src/main/java/kitchenpos/order/domain/OrderTableRepository.java
@@ -1,8 +1,9 @@
-package kitchenpos.ordertable.domain;
+package kitchenpos.order.domain;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.order.domain.OrderTable;
 
 public interface OrderTableRepository {
     OrderTable save(OrderTable orderTable);
