+++ src/main/java/kitchenpos/order/infrastructure/JpaOrderTableRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.ordertable.infrastructure;
+package kitchenpos.order.infrastructure;
 
 import java.util.UUID;
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.OrderTableRepository;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
